+++ SBBDesignSystemMobileSwiftUI/Views/SBBOnboardingWrapper/SBBOnboardingWrapperCardView.swift
@@ -4,7 +4,67 @@
 
 import SwiftUI
 
-public struct SBBCardView: View {
+@available(*, deprecated, message: "renamed to `SBBOnboardingWrapperCardView` for consistency. Parameter `nextCard` not needed anymore, `onNext` takes no parameter.")
+public struct SBBCardView<Card: Equatable>: View {
+    
+    private let image: Image
+    private let title: Text
+    private let text: Text
+    private let size: CGSize?
+    private let canBeTried: Bool
+    private let nextCard: Card?
+    private let onNext: (Card?) -> Void
+    @Binding private var showTrySheet: Bool
+    
+    private let contentWidthLandscape: CGFloat?
+    private let buttonWidthLandscape: CGFloat?
+    
+    private let scrollViewHeight: CGFloat?
+    
+    public init(image: Image, title: Text, text: Text, size: CGSize? = nil, showTrySheet: Binding<Bool>, nextCard: Card, onNext: @escaping (Card?) -> Void) {
+        self.image = image
+        self.title = title
+        self.text = text
+        self.size = size
+        self.canBeTried = true
+        self._showTrySheet = showTrySheet
+        self.nextCard = nextCard
+        self.onNext = onNext
+        
+        self.contentWidthLandscape = size?.width != nil ? 3 * size!.width / 4 : nil
+        self.buttonWidthLandscape = size?.width != nil ? size!.width / 4 : nil
+        self.scrollViewHeight = size?.width != nil ? size!.height - 54 - 16 : nil
+    }
+    
+    public init(image: Image, title: Text, text: Text, size: CGSize? = nil, nextCard: Card, onNext: @escaping (Card?) -> Void) {
+        self.image = image
+        self.title = title
+        self.text = text
+        self.size = size
+        self.canBeTried = false
+        self._showTrySheet = .constant(false)
+        self.nextCard = nextCard
+        self.onNext = onNext
+        
+        self.contentWidthLandscape = size?.width != nil ? 3 * size!.width / 4 : nil
+        self.buttonWidthLandscape = size?.width != nil ? size!.width / 4 : nil
+        self.scrollViewHeight = size?.width != nil ? size!.height - 54 - 16 : nil
+    }
+    
+    public var body: some View {
+        if canBeTried {
+            SBBOnboardingWrapperCardView(image: image, title: title, text: text, size: size, showTrySheet: $showTrySheet, onNext: {
+                onNext(nextCard)
+            })
+        } else {
+            SBBOnboardingWrapperCardView(image: image, title: title, text: text, size: size, onNext: {
+                onNext(nextCard)
+            })
+        }
+    }
+}
+
+public struct SBBOnboardingWrapperCardView: View {
     
     private let image: Image
     private let title: Text
@@ -231,9 +291,9 @@ public struct SBBCardView: View {
 }
 
 #Preview("Without try function") {
-    SBBCardView(image: Image("Onboarding_Card1"), title: Text("Title"), text: Text("This is the content of the card"), onNext: { })
+    SBBOnboardingWrapperCardView(image: Image("Onboarding_Card1"), title: Text("Title"), text: Text("This is the content of the card"), onNext: { })
 }
 
 #Preview("With try function") {
-    SBBCardView(image: Image("Onboarding_Card1"), title: Text("Title"), text: Text("This is the content of the card"), showTrySheet: .constant(false), onNext: { })
+    SBBOnboardingWrapperCardView(image: Image("Onboarding_Card1"), title: Text("Title"), text: Text("This is the content of the card"), showTrySheet: .constant(false), onNext: { })
 }