+++ SBBDesignSystemMobileSwiftUI/Views/SBBOnboardingWrapper/SBBOnboardingWrapperSettingsView.swift
@@ -1,0 +1,11 @@
+//
+// Copyright (c) Schweizerische Bundesbahnen SBB, 2025
+//  
+
+import SwiftUI
+
+struct SBBOnboardingWrapperSettingsView: View {
+    var body: some View {
+        VStack {}
+    }
+}